requirements.txt
Django==3.2
pytz==2023.3
django-rest-passwordreset
djangorestframework
psycopg2-binary
pyyaml
ujson
requests
django_filter
